Moved permissions middlewares to the modules where they are used
